+++ lib/BitrateTest.ts
@@ -1,7 +1,7 @@
 import { EventEmitter } from 'events';
 import { BYTES_KEEP_BUFFERED, MAX_NUMBER_PACKETS, MIN_BITRATE_THRESHOLD, TEST_PACKET } from './constants';
 import { DiagnosticError } from './errors/DiagnosticError';
-import { NetworkTiming, TimeMeasurement } from './timing';
+import { TimeMeasurement } from './types';
 import {
   getRTCIceCandidateStatsReport,
   RTCIceCandidateStats,
@@ -48,19 +48,6 @@ export declare interface BitrateTest {
     event: BitrateTest.Events.End,
     listener: (report: BitrateTest.Report) => any,
   ): this;
-
-  /**
-   * Raised when the test encounters a warning such as HighFirstPacketDuration.
-   * See [[BitrateTest.Warnings]] for more information.
-   * @param event [[BitrateTest.Events.Warning]].
-   * @param listener A callback with a [[BitrateTest.Warnings]] parameter.
-   * @returns This [[BitrateTest]] instance.
-   * @event
-   */
-  on(
-    event: BitrateTest.Events.Warning,
-    listener: (warning: BitrateTest.Warnings) => any,
-  ): this;
 }
 
 /**
@@ -82,6 +69,11 @@ export class BitrateTest extends EventEmitter {
   private _checkBitrateIntervalId: NodeJS.Timer | undefined;
 
   /**
+   * A timestamp of when the test ends.
+   */
+  private _endTime: number | null = null;
+
+  /**
    * Errors detected during the test
    */
   private _errors: DiagnosticError[] = [];
@@ -100,11 +92,6 @@ export class BitrateTest extends EventEmitter {
    * Last timestamp when the bytes received was checked
    */
   private _lastCheckedTimestamp: number = 0;
-
-  /**
-   * Network related timing for this test
-   */
-  private _networkTiming: NetworkTiming = {};
 
   /**
    * The options passed to [[BitrateTest]] constructor.
@@ -142,9 +129,10 @@ export class BitrateTest extends EventEmitter {
   private _sendDataIntervalId: NodeJS.Timer | undefined;
 
   /**
-   * Timing measurements for this test
+   * A timestamp of when the test starts. This is set during initialization of the test
+   * and not when the test succesfully starts.
    */
-  private _testTiming: TimeMeasurement = { start: 0 };
+  private _startTime: number;
 
   /**
    * Total number of bytes received by the receiver RTCPeerConnection
@@ -155,11 +143,6 @@ export class BitrateTest extends EventEmitter {
    * Bitrate (kbps) values collected during the test
    */
   private _values: number[] = [];
-
-  /**
-   * Warnings detected during the test
-   */
-  private _warnings: BitrateTest.Warnings[] = [];
 
   /**
    * Construct a [[BitrateTest]] instance. The test will start immediately.
@@ -181,6 +164,8 @@ export class BitrateTest extends EventEmitter {
 
     this._setupNetworkListeners(this._pcSender);
 
+    this._startTime = Date.now();
+
     // Return before starting the test to allow consumer
     // to listen and capture errors
     setTimeout(() => {
@@ -198,9 +183,7 @@ export class BitrateTest extends EventEmitter {
 
     this._pcSender.close();
     this._pcReceiver.close();
-
-    this._testTiming.end = Date.now();
-    this._testTiming.duration = this._testTiming.end - this._testTiming.start;
+    this._endTime = Date.now();
 
     this.emit(BitrateTest.Events.End, this._getReport());
   }
@@ -234,16 +217,20 @@ export class BitrateTest extends EventEmitter {
       .reduce((total: number, value: number) => total += value, 0) / this._values.length;
     averageBitrate = isNaN(averageBitrate) ? 0 : averageBitrate;
 
+    const testTiming: TimeMeasurement = { start: this._startTime };
+    if (this._endTime) {
+      testTiming.end = this._endTime;
+      testTiming.duration = this._endTime - this._startTime;
+    }
+
     const report: BitrateTest.Report = {
       averageBitrate,
       didPass: !this._errors.length && !!this._values.length && averageBitrate >= MIN_BITRATE_THRESHOLD,
       errors: this._errors,
       iceCandidateStats: this._iceCandidateStats,
-      networkTiming: this._networkTiming,
       testName: BitrateTest.testName,
-      testTiming: this._testTiming,
+      testTiming,
       values: this._values,
-      warnings: this._warnings,
     };
 
     if (this._selectedIceCandidatePairStats) {
@@ -251,18 +238,6 @@ export class BitrateTest extends EventEmitter {
     }
 
     return report;
-  }
-
-  /**
-   * Emit a warning if currentStatValue exceeds threshold value
-   * @param name
-   * @param currentStatValue
-   */
-  private _maybeEmitWarning(name: BitrateTest.Warnings, currentStatValue: Number): void {
-    if (currentStatValue > BitrateTest.WarningThresholds[name]) {
-      this._warnings.push(name);
-      this.emit(BitrateTest.Events.Warning, name);
-    }
   }
 
   /**
@@ -298,14 +273,6 @@ export class BitrateTest extends EventEmitter {
    */
   private _onMessageReceived(event: MessageEvent) {
     this._totalBytesReceived += event.data.length;
-
-    if (!this._networkTiming.firstPacket) {
-      this._networkTiming.firstPacket = Date.now();
-      this._maybeEmitWarning(
-        BitrateTest.Warnings.HighFirstPacketDuration,
-        this._networkTiming.firstPacket - this._testTiming.start,
-      );
-    }
   }
 
   /**
@@ -373,36 +340,8 @@ export class BitrateTest extends EventEmitter {
    * @param pc
    */
   private _setupNetworkListeners(pc: RTCPeerConnection) {
-    // PeerConnection state
-    pc.onconnectionstatechange = () => {
-      this._networkTiming.peerConnection = this._networkTiming.peerConnection || { start: 0 };
-
-      if (pc.connectionState === 'connecting') {
-        this._networkTiming.peerConnection.start = Date.now();
-      } else if (pc.connectionState === 'connected') {
-        this._networkTiming.peerConnection.end = Date.now();
-
-        const { start, end } = this._networkTiming.peerConnection;
-        const duration = end - start;
-        this._networkTiming.peerConnection.duration = duration;
-        this._maybeEmitWarning(BitrateTest.Warnings.HighPcConnectDuration, duration);
-      }
-    };
-
-    // ICE Connection state
     pc.oniceconnectionstatechange = () => {
-      this._networkTiming.ice = this._networkTiming.ice || { start: 0 };
-
-      if (pc.iceConnectionState === 'checking') {
-        this._networkTiming.ice.start = Date.now();
-      } else if (pc.iceConnectionState === 'connected') {
-        this._networkTiming.ice.end = Date.now();
-
-        const { start, end } = this._networkTiming.ice;
-        const duration = end - start;
-        this._networkTiming.ice.duration = duration;
-        this._maybeEmitWarning(BitrateTest.Warnings.HighIceConnectDuration, duration);
-
+      if (pc.iceConnectionState === 'connected') {
         (this._options.getRTCIceCandidateStatsReport || getRTCIceCandidateStatsReport)(this._pcSender)
           .then((statsReport: RTCIceCandidateStatsReport) => {
             this._iceCandidateStats = statsReport.iceCandidateStats;
@@ -419,8 +358,6 @@ export class BitrateTest extends EventEmitter {
    * Starts the test.
    */
   private _startTest(): void {
-    this._testTiming.start = Date.now();
-
     if (!this._rtcConfiguration.iceServers) {
       return this._onError('No iceServers found', undefined);
     }
@@ -443,39 +380,7 @@ export namespace BitrateTest {
     Bitrate = 'bitrate',
     End = 'end',
     Error = 'error',
-    Warning = 'warning',
   }
-
-  /**
-   * Possible warnings that a [[BitrateTest]] might emit. See [[BitrateTest.on]].
-   */
-  export enum Warnings {
-    /**
-     * Raised when [[NetworkTiming.firstPacket]] took more than 1400ms to arrive to the remote RTCPeerConnection.
-     * The duration is measured from the [[BitrateTest.Report]]'s testTiming.start up to [[NetworkTiming.firstPacket]]
-     */
-    HighFirstPacketDuration = 'high-first-packet-duration',
-
-    /**
-     * Raised when [[NetworkTiming.ice]] connection took more than 300ms to establish
-     */
-    HighIceConnectDuration = 'high-ice-connect-duration',
-
-    /**
-     * Raised when [[NetworkTiming.peerConnection]] took more than 1000ms to establish
-     */
-    HighPcConnectDuration = 'high-pc-connect-duration',
-  }
-
-  /**
-   * Thresholds used for determining when to raise a warning. See [[BitrateTest.Warnings]]
-   * @internalapi
-   */
-  export const WarningThresholds: Record<Warnings, Number> = {
-    [Warnings.HighFirstPacketDuration]: 1400,
-    [Warnings.HighIceConnectDuration]: 300,
-    [Warnings.HighPcConnectDuration]: 1000,
-  };
 
   /**
    * Options that may be passed to [[BitrateTest]] constructor for internal testing.
@@ -557,11 +462,6 @@ export namespace BitrateTest {
     iceCandidateStats: RTCIceCandidateStats[];
 
     /**
-     * Network related time measurements.
-     */
-    networkTiming: NetworkTiming;
-
-    /**
      * A WebRTC stats for the ICE candidate pair used to connect to media, if candidates were selected.
      */
     selectedIceCandidatePairStats?: RTCSelectedIceCandidatePairStats;
@@ -580,11 +480,6 @@ export namespace BitrateTest {
      * Bitrate values collected during the test.
      */
     values: number[];
-
-    /**
-     * Any warnings that occurred during the test.
-     */
-    warnings: Warnings[];
   }
 }